+++ expense-manage/src/main/java/com/porejct/expensemanage/domain/budget/service/BudgetService.java
@@ -4,6 +4,7 @@ import static com.porejct.expensemanage.domain.budget.exception.BudgetExceptionC
 
 import com.porejct.expensemanage.commone.exception.BusinessLogicException;
 import com.porejct.expensemanage.domain.budget.dto.BudgetMapper;
+import com.porejct.expensemanage.domain.budget.dto.request.PostBudgetRequest;
 import com.porejct.expensemanage.domain.budget.dto.response.BudgetIdResponse;
 import com.porejct.expensemanage.domain.budget.entity.Budget;
 import com.porejct.expensemanage.domain.budget.repository.BudgetRepository;
