+++ expense-manage/src/test/java/com/porejct/expensemanage/domain/budget/controller/BudgetControllerTest.java
@@ -10,7 +10,7 @@ import com.porejct.expensemanage.commone.security.annotation.WithMockCustomUser;
 import com.porejct.expensemanage.commone.security.config.AuthTestConfig;
 import com.porejct.expensemanage.domain.budget.mock.BudgetMock;
 import com.porejct.expensemanage.domain.budget.service.BudgetService;
-import com.porejct.expensemanage.domain.budget.service.PostBudgetRequest;
+import com.porejct.expensemanage.domain.budget.dto.request.PostBudgetRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.BDDMockito;
@@ -19,7 +19,6 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
-import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
